+++ addon/adapters/odata.js
@@ -4,6 +4,8 @@ import DS from 'ember-data';
 import SnapshotTransform from '../utils/snapshot-transform';
 import ODataQueryAdapter from '../query/odata-adapter';
 import { capitalize, camelize } from '../utils/string-functions';
+import generateUniqueId from '../utils/generate-unique-id';
+import { getResponseMeta, getBatchResponses, parseBatchResponse } from '../utils/batch-queries';
 
 const { getOwner } = Ember;
 
@@ -19,6 +21,7 @@ const { getOwner } = Ember;
  */
 export default DS.RESTAdapter.extend({
   headers: {
+    'OData-Version': '4.0',
     Prefer: 'return=representation'
   },
 
@@ -42,6 +45,10 @@ export default DS.RESTAdapter.extend({
     @return {Promise}
   */
   query(store, type, query) {
+    if (!this.store) {
+      this.store = store;
+    }
+
     let url = this._buildURL(query.modelName);
     let builder = new ODataQueryAdapter(url, store);
     let data = builder.getODataQuery(query);
@@ -241,7 +248,7 @@ export default DS.RESTAdapter.extend({
       url = `${config.APP.backendUrls.api}`;
     }
 
-    const resultUrl =  `${url}/${actionName}`;
+    const resultUrl = `${url}/${actionName}`;
 
     return resultUrl;
   },
@@ -265,7 +272,7 @@ export default DS.RESTAdapter.extend({
     const resultUrl = this.generateFunctionUrl(functionName, params, url);
     return this._callAjax(
       { url: resultUrl, method: 'GET', xhrFields: Ember.isNone(fields) ? {} : fields },
-        store, modelName, successCallback, failCallback, alwaysCallback);
+      store, modelName, successCallback, failCallback, alwaysCallback);
   },
 
   /**
@@ -286,7 +293,7 @@ export default DS.RESTAdapter.extend({
     const resultUrl = this.generateFunctionUrl(functionName, params, url);
     return this._callAjax(
       { url: resultUrl, method: 'GET', xhrFields: Ember.isNone(fields) ? {} : fields },
-        null, null, successCallback, failCallback, alwaysCallback);
+      null, null, successCallback, failCallback, alwaysCallback);
 
   },
 
@@ -308,7 +315,8 @@ export default DS.RESTAdapter.extend({
   callEmberOdataAction(actionName, data, url, fields, store, modelName, successCallback, failCallback, alwaysCallback) {
     const resultUrl = this.generateActionUrl(actionName, data, url);
     return this._callAjax(
-      { data: JSON.stringify(data),
+      {
+        data: JSON.stringify(data),
         url: resultUrl,
         method: 'POST',
         contentType: 'application/json; charset=utf-8',
@@ -334,13 +342,177 @@ export default DS.RESTAdapter.extend({
   callAction(actionName, data, url, fields, successCallback, failCallback, alwaysCallback) {
     const resultUrl = this.generateActionUrl(actionName, data, url);
     return this._callAjax(
-      { data: JSON.stringify(data),
+      {
+        data: JSON.stringify(data),
         url: resultUrl,
         method: 'POST',
         contentType: 'application/json; charset=utf-8',
         dataType: 'json',
         xhrFields: Ember.isNone(fields) ? {} : fields
       }, null, null, successCallback, failCallback, alwaysCallback);
+  },
+
+  /**
+    A method to send batch update, create or delete models in single transaction.
+
+    The array which fulfilled the promise may contain the following values:
+    - `new model object` - for created records.
+    - `same model object` - for updated or unaltered records.
+    - `null` - for deleted records.
+
+    @method batchUpdate
+    @param {DS.Store} store The store.
+    @param {DS.Model[]|DS.Model} models Is array of models or single model for batch update.
+    @return {Promise} A promise that fulfilled with an array of models in the new state.
+  */
+  batchUpdate(store, models) {
+    if (Ember.isEmpty(models)) {
+      throw new Error('The models is empty.');
+    }
+
+    models = Ember.A(models);
+
+    const boundary = `batch_${generateUniqueId()}`;
+
+    let requestBody = '--' + boundary + '\r\n';
+
+    const changeSetBoundary = `changeset_${generateUniqueId()}`;
+    requestBody += 'Content-Type: multipart/mixed;boundary=' + changeSetBoundary + '\r\n\r\n';
+
+    let contentId = 0;
+    const modelsByContentId = {};
+    models.forEach((model) => {
+      let modelDirtyType = model.get('dirtyType');
+
+      if (!modelDirtyType) {
+        return;
+      }
+
+      requestBody += '--' + changeSetBoundary + '\r\n';
+      requestBody += 'Content-Type: application/http\r\n';
+      requestBody += 'Content-Transfer-Encoding: binary\r\n';
+
+      contentId++;
+      requestBody += 'Content-ID: ' + contentId + '\r\n\r\n';
+      modelsByContentId[contentId] = model;
+
+      const skipUnchangedAttrs = true;
+      const snapshot = model._createSnapshot();
+      SnapshotTransform.transformForSerialize(snapshot, skipUnchangedAttrs);
+
+      let modelHttpMethod = 'POST';
+      switch (modelDirtyType) {
+        case 'created':
+          modelHttpMethod = 'POST';
+          break;
+
+        case 'updated':
+          modelHttpMethod = skipUnchangedAttrs ? 'PATCH' : 'PUT';
+          break;
+
+        case 'deleted':
+          modelHttpMethod = 'DELETE';
+          break;
+
+        default:
+          throw new Error(`Unknown requestType: ${modelDirtyType}`);
+      }
+
+      if (!this.store) {
+        this.store = store;
+      }
+
+      const modelUrl =  this._buildURL(snapshot.type.modelName, model.get('id'));
+
+      requestBody += modelHttpMethod + ' ' + modelUrl + ' HTTP/1.1\r\n';
+      requestBody += 'Content-Type: application/json;type=entry\r\n';
+      requestBody += 'Prefer: return=representation\r\n\r\n';
+
+      // Don't need to send any data for deleting.
+      if (modelDirtyType !== 'deleted') {
+        const serializer = store.serializerFor(snapshot.type.modelName);
+        const data = {};
+        serializer.serializeIntoHash(data, snapshot.type, snapshot);
+        requestBody += JSON.stringify(data) + '\r\n';
+      }
+    });
+
+    requestBody += '--' + changeSetBoundary + '--\r\n';
+
+    requestBody += '\r\n--' + boundary + '--';
+
+    const url = `${this._buildURL()}/$batch`;
+
+    const headers = this.get('headers');
+    headers['Content-Type'] = `multipart/mixed;boundary=${boundary}`;
+    const httpMethod = 'POST';
+
+    const options = {
+      method: httpMethod,
+      headers,
+      dataType: 'text',
+      data: requestBody,
+    };
+
+    return new Ember.RSVP.Promise((resolve, reject) => Ember.$.ajax(url, options).done((response, statusText, xhr) => {
+      const meta = getResponseMeta(xhr.getResponseHeader('Content-Type'));
+      if (meta.contentType !== 'multipart/mixed') {
+        return reject(new Error('Invalid response type.'));
+      }
+
+      const responses = getBatchResponses(response, meta.boundary);
+      if (responses.length !== 1) {
+        return reject(new Error('Invalid number of responses.'));
+      }
+
+      const { contentType, changesets } = parseBatchResponse(responses[0]);
+      if (contentType !== 'multipart/mixed') {
+        return reject(new Error('Invalid response type.'));
+      }
+
+      if (changesets.length === 1 && !this.isSuccess(changesets[0].meta.status)) {
+        return reject(new Error('Invalid response.'));
+      }
+
+      const result = [];
+      for (const model of models) {
+        const changeset = changesets.find(c => modelsByContentId[c.contentID] === model);
+        const modelClass = store.modelFor(model.constructor.modelName);
+        const serializer = store.serializerFor(modelClass.modelName);
+        switch (model.get('dirtyType')) {
+          case 'created':
+            if (changeset.meta.status !== 201) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            Ember.run(store, store.unloadRecord, model);
+            result.push(Ember.run(store, store.push, serializer.normalize(modelClass, changeset.body)));
+            break;
+
+          case 'updated':
+            if (changeset.meta.status !== 200) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            result.push(Ember.run(store, store.push, serializer.normalize(modelClass, changeset.body)));
+            break;
+
+          case 'deleted':
+            if (changeset.meta.status !== 204) {
+              return reject(new Error(`Invalid response status: ${changeset.meta.status}.`));
+            }
+
+            Ember.run(store, store.unloadRecord, model);
+            result.push(null);
+            break;
+
+          default:
+            result.push(model);
+        }
+      }
+
+      resolve(result);
+    }).fail(reject));
   },
 
   /**
@@ -359,7 +531,7 @@ export default DS.RESTAdapter.extend({
     Ember.assert('Params must be Object!', typeof params === 'object');
     Ember.assert('params.method or params.url is not defined.', !(Ember.isNone(params.method) || Ember.isNone(params.url)));
 
-    return new Ember.RSVP.Promise(function(resolve, reject) {
+    return new Ember.RSVP.Promise(function (resolve, reject) {
       Ember.$.ajax(params).done((msg) => {
         if (!Ember.isNone(store) && !Ember.isNone(modelname)) {
           const normalizedRecords = { data: [] };
@@ -373,7 +545,7 @@ export default DS.RESTAdapter.extend({
           if (typeof successCallback.then === 'function') {
             if (!Ember.isNone(alwaysCallback)) {
               if (typeof alwaysCallback.then === 'function') {
-                successCallback(msg).then(() => {alwaysCallback(msg).then(resolve(msg));});
+                successCallback(msg).then(() => { alwaysCallback(msg).then(resolve(msg)); });
               } else {
                 successCallback(msg).then(alwaysCallback(msg)).then(resolve(msg));
               }
@@ -405,12 +577,12 @@ export default DS.RESTAdapter.extend({
             resolve(msg);
           }
         }
-      }).fail((msg)=> {
+      }).fail((msg) => {
         if (!Ember.isNone(failCallback)) {
           if (typeof failCallback.then === 'function') {
             if (!Ember.isNone(alwaysCallback)) {
               if (typeof alwaysCallback.then === 'function') {
-                failCallback(msg).then(() => {alwaysCallback(msg).then(reject(msg));});
+                failCallback(msg).then(() => { alwaysCallback(msg).then(reject(msg)); });
               } else {
                 failCallback(msg).then(alwaysCallback(msg)).then(reject(msg));
               }
@@ -498,6 +670,11 @@ export default DS.RESTAdapter.extend({
    */
   _appendIdToURL(id, url, modelName) {
     let encId = encodeURIComponent(id);
+
+    if (!this.store) {
+      throw new Error('No store.');
+    }
+
     let model = this.store.modelFor(modelName);
     if (model.idType === 'string') {
       encId = `'${encId}'`;
@@ -520,8 +697,12 @@ export default DS.RESTAdapter.extend({
   },
 
   deleteAllRecords(store, modelName, filter) {
+    if (!this.store) {
+      this.store = store;
+    }
+
     let url = this._buildURL(modelName);
-    let pathName  = this.pathForType(modelName);
+    let pathName = this.pathForType(modelName);
     let builder = new ODataQueryAdapter(url, store);
     let filterVelue = builder._buildODataFilters(filter);
     let filterQuery = !Ember.isNone(filterVelue) ? '$filter=' + filterVelue : '';
@@ -571,7 +752,7 @@ export default DS.RESTAdapter.extend({
       serializer.serializeIntoHash(data, type, snapshot);
     }
 
-    return this.ajax(url, httpMethod, { data: data }).then(function(response) {
+    return this.ajax(url, httpMethod, { data: data }).then(function (response) {
       return response;
     });
   }