+++ tests/unit/CRUD/odata/execute-odata-test.js
@@ -11,17 +11,10 @@ export default function executeTest(testName, callback) {
   if (config.APP.testODataService) {
     let baseUrl;
     if (config.APP.testODataServiceURL.indexOf('http') >= 0) {
-      let index = config.APP.testODataServiceURL.lastIndexOf('/');
-      if (index !== config.APP.testODataServiceURL.length - 1) {
-        baseUrl = config.APP.testODataServiceURL + '/';
-      } else {
-        baseUrl = config.APP.testODataServiceURL;
-      }
+      baseUrl = config.APP.testODataServiceURL;
     } else {
-      baseUrl = 'http://rtc-web:8081/odatatmp/';
+      baseUrl = 'http://localhost:6500/odata';
     }
-
-    baseUrl += 'ember' + Math.floor(Math.random() * 9999);
 
     const app = startApp();
     const store = app.__container__.lookup('service:store');