+++ tests/unit/CRUD/odata/odata-batch-updating-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-odata-test';
+import batchUpdating from '../base/base-batch-updating-test';
+
+executeTest('batchUpdating', (store, assert) => {
+  batchUpdating(store, assert);
+});